generate openapi and fix nit comment

+++ go/apps/api/routes/v2_identities_create_identity/200_test.go
@@ -92,7 +92,7 @@ func TestCreateIdentitySuccessfully(t *testing.T) {
 		require.Len(t, rateLimits, 1)
 		require.Equal(t, rateLimits[0].Name, "Requests")
 		require.Equal(t, rateLimits[0].Limit, int32(15))
-		require.Equal(t, rateLimits[0].Duration, int64((time.Minute * 15).Milliseconds()))
+		require.Equal(t, rateLimits[0].Duration, (time.Minute * 15).Milliseconds())
 	})
 
 	// Test creating a identity with no other information
